% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape-addin.R
\name{addin_reshape}
\alias{addin_reshape}
\title{Reshape expressions longer or wider}
\usage{
addin_reshape()
}
\description{
\code{addin_reshape()} lets you cycle between different shapes of
function calls. For instance, reshaping transforms code from wide to long
shape and vice versa:

\if{html}{\out{<div class="sourceCode">}}\preformatted{list(a, b, c)

list(
  a,
  b,
  c
)
}\if{html}{\out{</div>}}

Note that for function definitions, \code{addin_reshape()} cycles through two
different long shapes. The traditional L form uses more horizontal space
whereas the flat form uses less horizontal space and the arguments are
always aligned at single indent:

\if{html}{\out{<div class="sourceCode">}}\preformatted{foo <- function(a, b, c) \{
  NULL
\}

foo <- function(a,
                b,
                c) \{
  NULL
\}

foo <- function(
  a,
  b,
  c
) \{
  NULL
\}
}\if{html}{\out{</div>}}
}
