% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegrip-package.R
\docType{package}
\name{codegrip-package}
\alias{codegrip}
\alias{codegrip-package}
\title{codegrip: Syntax-Based Editing and Navigation of R Code}
\description{
Provides addins for reshaping R code and navigating across syntactic constructs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lionel-/codegrip}
  \item Report bugs at \url{https://github.com/lionel-/codegrip/issues}
}

}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@posit.co}

Other contributors:
\itemize{
  \item Posit PBC [copyright holder, funder]
}

}
\keyword{internal}
