% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape.R
\name{reshape_info}
\alias{reshape_info}
\title{Prepare Reshape Information}
\usage{
reshape_info(line, col, ..., info, to = NULL)
}
\arguments{
\item{line, col}{\code{integer} The location of the cursor as the focus of the
reshape action.}

\item{...}{Arguments unused}

\item{info}{\code{list} of named elements \code{file} (\code{string}), \code{lines}
(\code{character()}) and \code{xml} (\code{\link[xml2:xml_new_document]{xml2::xml_new_document()}}), as produced
using \code{parse_info()}.}

\item{to}{\code{string} An optional hint for the reshape action. Expecting one of
\code{"wide"} (single line call), \code{"L"} (indented to the call left parenthesis)
or \code{"long"} (arguments indented by one indentation).}
}
\description{
Prior to performing an editor-specific operation, prepare minimal
information needed to perform the action.
}
