% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codegrip-package.R
\docType{package}
\name{codegrip-package}
\alias{codegrip}
\alias{codegrip-package}
\title{codegrip: Shape R code}
\description{
Shape R code.
}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@posit.co}

Other contributors:
\itemize{
  \item Posit PBC [copyright holder, funder]
}

}
\keyword{internal}
